+++ src/imio/smartweb/common/sharing/__init__.py
@@ -1,0 +1,1 @@
+import imio.smartweb.common.sharing.permissions  # NOQA